fix(research-pipe): handle case when chunk is of type bytes
The implementation of arXiv and Web Research MCTS Pipe currently assumes it gets a chunk of type string, however this can at times also be of type bytes. Crashing the research pipe and preventing it from returning results.

This fix adds validation of the type of chunk, leaving the situation unchanged if it was already a string. In case the type of chunk is bytes, we will decode it as if it's UTF-8 and convert it into a string.

Tested with OpenWebUI v0.6.34.